tests/requirements.txt
pytest>=7.4.0
requests>=2.31.0
pytest-docker>=2.0.0
psycopg2-binary>=2.9.9



